% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package.R
\name{load_package}
\alias{load_package}
\title{Loads and installs package from CRAN or github}
\usage{
load_package(package_names)
}
\arguments{
\item{package_name}{Character vector with package name to install.}

\item{repo}{Character vector with github repo name to install. Default is
\code{NULL}, which forces a CRAN install.}

\item{branch}{Character vector with branch name to install. Default is
\code{NULL} which install from master if \code{repo} is not \code{NULL}.}
}
\value{
something
}
\description{
Loads and installs package from CRAN or github
}
\examples{
load_package("beepr")
}
\author{
Pippo Baudo, Giovanni Laudanno, Pedro Neves
}
