% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hello_world.R
\name{print_hello_world}
\alias{print_hello_world}
\title{Print hello world}
\usage{
print_hello_world()
}
\value{
Character vector: "Hello world!"
}
\description{
Print hello world
}
\examples{
print_hello_world()
}
